adroid host 1

+++ android host/telegram_bot_control.py
@@ -102,8 +102,11 @@ def get_ip():
 def take_screenshot():
     try:
         subprocess.run(['termux-media-capture', '--screenshot', '/sdcard/screenshot.png'], check=True)
-        bot.send_photo(GROUP_ID, open('/sdcard/screenshot.png', 'rb'), caption="📸 Скриншот сделан")
-        os.remove('/sdcard/screenshot.png')
+        if os.path.isfile('/sdcard/screenshot.png') and os.path.getsize('/sdcard/screenshot.png') > 0:
+            bot.send_photo(GROUP_ID, open('/sdcard/screenshot.png', 'rb'), caption="📸 Скриншот сделан")
+            os.remove('/sdcard/screenshot.png')
+        else:
+            bot.send_message(GROUP_ID, "📸 Ошибка: Скриншот не создан или пустой.")
     except Exception as e:
         bot.send_message(GROUP_ID, f"📸 Ошибка при создании скриншота: {e}")
 
@@ -124,18 +127,21 @@ def get_sysinfo():
 def take_webcam_screenshot():
     try:
         subprocess.run(['termux-camera-photo', '-c', '0', '/sdcard/webcam.jpg'], check=True)
-        bot.send_photo(GROUP_ID, open('/sdcard/webcam.jpg', 'rb'), caption="📷 Снимок с камеры")
-        os.remove('/sdcard/webcam.jpg')
+        if os.path.isfile('/sdcard/webcam.jpg') and os.path.getsize('/sdcard/webcam.jpg') > 0:
+            bot.send_photo(GROUP_ID, open('/sdcard/webcam.jpg', 'rb'), caption="📷 Снимок с камеры")
+            os.remove('/sdcard/webcam.jpg')
+        else:
+            bot.send_message(GROUP_ID, "📷 Ошибка: Снимок камеры не создан или пустой.")
     except Exception as e:
         bot.send_message(GROUP_ID, f"📷 Ошибка при использовании камеры: {e}")
 
 def shutdown_device():
-    bot.send_message(GROUP_ID, "⏻ Выключение недоступно на Android без root.")
+    bot.send_message(GROUP_ID, "⏻ Выключение недоступно на Android без root")
 
 def wake_on_lan():
     mac_address = '5A:57:F7:C7:D7:0A'
     send_magic_packet(mac_address)
-    bot.send_message(GROUP_ID, "🔌 Отправлен сигнал Wake-on-LAN.")
+    bot.send_message(GROUP_ID, "🔌 Отправлен сигнал Wake-on-LAN")
 
 def execute_cmd(command):
     try:
@@ -160,12 +166,12 @@ def set_wallpaper_from_url(url):
         bot.send_message(GROUP_ID, f"🖼️ Ошибка при установке обоев: {e}")
 
 def set_wallpaper_from_file(file_path):
-    try:
-        subprocess.run(['termux-wallpaper', '-f', file_path], check=True)
-        bot.send_message(GROUP_ID, "🖼️ Обои установлены!")
-        os.remove(file_path)
-    except Exception as e:
-        bot.send_message(GROUP_ID, f"🖼️ Ошибка при установке файла: {e}")
+        try:
+            subprocess.run(['termux-wallpaper', '-f', file_path], check=True)
+            bot.send_message(GROUP_ID, "🖼️ Обои установлены!")
+            os.remove(file_path)
+        except Exception as e:
+            bot.send_message(GROUP_ID, f"🖼️ Ошибка при установке файла: {e}")
 
 def open_url(url):
     try:
